Vinyl Record GUI
This was the hardest program I've worked on so far in my progress of learning C#. I had a lot of fun doing it. It is not completely done, I want to go over the code as I learn more efficient ways of C# to improve my code. Eventually I would love to have this available to where it does not need a SQL database. All in due time! For now, this uses an existing SQL Database, and configure Windows Forms in Visual Studio.